Add zstd support to `conan upload`
This change adds zstd support to `conan upload` in the form of a new
command-line parameter `--binary-compression-format`. When set to zstd, the
uploader calls into `tar` to compress using zstd. This is necessary because
zstd itself doesn't support compressing anything other than files. So tar+zstd
is the way to compress multiple files at once.