Add support for get() with .tar.zst files
My previous changes already supported upload, download, and extract of .tar.zst
package files. But it makes sense to also support this for get() inside of Conan
recipes. This change does that.

As far as I can tell, get() doesn't share untar code with the core Conan package
download logic, so I had to duplicate a bit of logic to do the extraction and
decompress of the .tar.zst file. If there is a better place to put this where
both modules can access, I can refactor in a follow-up change.